Define NetworkControllerSpecificInterface for functions that have different implementations based on the controller implementation (these are the functions that had 2 separate branches with "if bnc.IsSecondary()" condition). Move BaseSecondaryNetworkController definition to base_network_controller_secondary.go
